Move integration tests too
